修改实体字段

+++ models/SingerSongMapping.js
@@ -3,12 +3,12 @@ const db = require('../db/db');
 module.exports = db.defineModel('singer_song_mapping', {
   // 歌曲编号
   songId: {
-    type: db.INTEGER(11),
+    type: db.BIGINT,
     field: 'song_id'
   },
   // 歌手编号
   singerId: {
-    type: db.INTEGER(11),
+    type: db.BIGINT,
     field: 'singer_id'
   }
 })
